% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_act.R
\name{.bias_act_ref}
\alias{.bias_act_ref}
\title{Slow reference implementation of \code{bias_act()} using standard torch ops.}
\usage{
.bias_act_ref(
  x,
  b = NULL,
  dim = 1,
  spec = list(name = "linear", func = function(x, ...) x, def_alpha = 0, def_gain = 1,
    cuda_idx = 1, ref = "", has_2nd_grad = FALSE),
  alpha = NULL,
  gain = NULL,
  clamp = NULL
)
}
\description{
Slow reference implementation of \code{bias_act()} using standard torch ops.
}
