% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upfirdn2d.R
\name{filter2d}
\alias{filter2d}
\title{Filter a batch of 2D images using the given 2D FIR filter.}
\usage{
filter2d(
  x,
  f,
  padding = 0,
  flip_filter = FALSE,
  gain = 1,
  impl = if (cuda_is_available()) "cuda" else "ref"
)
}
\arguments{
\item{x}{Float32/float64/float16 input tensor of the shape
\code{c(batch_size, num_channels, in_height, in_width)}.}

\item{f}{Float32 FIR filter of the shape
\code{c(filter_height, filter_width)} (non-separable),
\code{filter_taps} (separable), or \code{NULL} (identity).}

\item{padding}{Padding with respect to the output. Can be a single number or a
vector \code{c(x, y)} or \code{c(x_before, x_after, y_before, y_after)}
(default: 0).}

\item{flip_filter}{\code{FALSE} = convolution, \code{TRUE} = correlation (default: \code{FALSE}).}

\item{gain}{Overall scaling factor for signal magnitude (default: 1).}

\item{impl}{Implementation to use. Can be \code{'ref'} or \code{'cuda'}
(default: \code{'cuda'} if \code{torch::cuda_is_available() == TRUE}, \code{'ref'} otherwise).}
}
\value{
Tensor of the shape \code{c(batch_size, num_channels, out_height, out_width)}.
}
\description{
By default, the result is padded so that its shape matches the input.
User-specified padding is applied on top of that, with negative values
indicating cropping. Pixels outside the image are assumed to be zero.
}
