% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{modulated_conv2d}
\alias{modulated_conv2d}
\title{Modulated Convolution}
\usage{
modulated_conv2d(x, w, s, demodulate = TRUE, padding = 0, input_gain = NULL)
}
\arguments{
\item{input_gain}{}
}
\value{

}
\description{
Modulated Convolution
}
