% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upfirdn2d.R
\name{setup_filter}
\alias{setup_filter}
\title{Convenience function to setup 2D FIR filter for \code{\link[=upfirdn2d]{upfirdn2d()}}.}
\usage{
setup_filter(
  f,
  device = torch_device("cpu"),
  normalize = TRUE,
  flip_filter = FALSE,
  gain = 1,
  separable = NULL
)
}
\arguments{
\item{f}{Torch tensor, or R array of the shape: \code{c(filter_height, filter_width)} (non-separable),
\code{filter_taps} (separable), \code{interger()} (impulse), or \code{NULL} (identity).}

\item{device}{Result device (default: cpu).}

\item{normalize}{Normalize the filter so that it retains the magnitude
for constant input signal (DC)? (default: TRUE).}

\item{flip_filter}{Flip the filter? (default: FALSE).}

\item{gain}{Overall scaling factor for signal magnitude (default: 1).}

\item{separable}{Return a separable filter? (default: select automatically).}
}
\value{
Float32 tensor of the shape: \code{c(filter_height, filter_width)} (non-separable) or
\code{filter_taps} (separable).
}
\description{
Convenience function to setup 2D FIR filter for \code{\link[=upfirdn2d]{upfirdn2d()}}.
}
