% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylegan_models.R
\name{sgr_get_model}
\alias{sgr_get_model}
\title{Get a pretrained StyleGAN3 model}
\usage{
sgr_get_model(
  model_name = c("afhqv2"),
  flavour = c("r", "t"),
  res = c("512", "1024"),
  dir = rappdirs::user_data_dir("styleganr"),
  device = c("cpu", "cuda")
)
}
\arguments{
\item{model_name}{Name of the model}

\item{flavour}{The flavour of the model: \code{'r'} for rotation-invariant, \code{'t'} for
translation invariant.}

\item{res}{The resolution of the model.}

\item{dir}{Directory to save the model to. By default it will be saved to your user
data directory.}

\item{device}{Device to load the model onto: \code{'cpu'} or \code{'gpu'}}
}
\value{
A \code{\link[torch:nn_module]{torch::nn_module()}} object.
}
\description{
The model will be downloaded to a local directory if you do not have it already.
otherwise the locally cached version will be used.
}
