% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleganr-package.R
\docType{package}
\name{styleganr-package}
\alias{styleganr}
\alias{styleganr-package}
\title{styleganr: Generate From and Train StyleGAN3 Models in R}
\description{
Provides R versions of StyleGAN3 models using the `torch` package for R. This let's user's generate images from pretrained model (even trained in Python), an even to train your own StyleGAN3, all in R (this requires one or more powerful GPUs).
}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/0000-0003-0846-2819}{ORCID}) [copyright holder]

}
\keyword{internal}
